+++ src/oihana/files/openssl/OpenSSLFileEncryption.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace oihana\files;
+namespace oihana\files\openssl;
 
 use Exception;
 use InvalidArgumentException;